modifications

+++ index.js
@@ -2,9 +2,10 @@ let myLeads = []
 const inputEl = document.getElementById("input-el")
 const inputBtn = document.getElementById("input-btn")
 const ulEl = document.getElementById("ul-el")
-const deleteBtn = document.getElementById("delete-btn")
 const tabBtn = document.getElementById("tab-btn")
+const selectBtn = document.getElementById("select-btn")
 
+const deleteBtn = document.getElementById("delete-btn")
 
 const leadsFromLocalStorage = JSON.parse(localStorage.getItem("myLeads"))
 
@@ -21,11 +22,15 @@ tabBtn.addEventListener("click", function() {
   })
 })
 
+
 function render(leads) {
   let listItems = ""
   for(let i = 0; i < leads.length; i ++) {
     listItems += `
     <li>
+      <button class="deleteOne-btn">
+        <i class="far fa-trash-alt"></i>
+      </button>
       <a href= '${leads[i]}' target='_blank'>
         ${leads[i]}
       </a>
@@ -35,15 +40,21 @@ function render(leads) {
   ulEl.innerHTML = listItems
 }
 
-deleteBtn.addEventListener("dblclick", function() {
-  localStorage.clear()
-  myLeads = []
-  render(myLeads)
+deleteBtn.addEventListener("click", function() {
+  if (window.confirm('Are you sure?')) {
+    localStorage.clear()
+    myLeads = []
+    render(myLeads)
+  }
 })
 
 inputBtn.addEventListener("click", function() {
-  myLeads.push(inputEl.value)
-  inputEl.value = ""
-  localStorage.setItem("myLeads", JSON.stringify(myLeads))
-  render(myLeads)
+  if (inputEl.value !== '' && !myLeads.includes(inputEl.value)) {
+    myLeads.push(inputEl.value)
+    inputEl.value = ""
+    localStorage.setItem("myLeads", JSON.stringify(myLeads))
+    render(myLeads)
+  } else {
+    alert("Please enter valid input")
+  }
 })
